refer/requirements.txt
fastapi
uvicorn
sqlalchemy
google-genai
googlemaps
python-dotenv
requests
pydantic